Channels do not always work as expected due to URL paremeters and undefined meta actions
When the channel is undefined, the config URL is used.
This config URL may contain query parameters, preventing string matching against channels.

When `meta.action` is not defined, it append to the string as `null`.
Avoid this behavior, in fact, if `meta.action` is not defined, do not operate any of the block of code using it.