Enable Verdaccio caching and storing in memory by default.
The expected default behavior is to run in docker.
It makes more sense to use local disk caching rather not caching and running from memory.

Enable caching of npmjs registry.
Set a max cache expiration large enough to always be within the sprint cycle but small enough to eventually clean itself up.

Disable storing memory by default.
The docker can always be pruned or deleted.
Using this in docker reduces some of the benefits of running from memory in terms of easier maintenance.