Support all known pattern functions and conditionally handle 'to'.
The `to()` function is needed and is unavailable.

Add all known functions rather than add the functions and require a new release every time a function is needed.

For the case of `to`, only call `resolve()` if `to` is not a function.

see: https://www.npmjs.com/package/copy-webpack-plugin#user-content-patterns