Issue 217: Weaver authentication is making unsafe assumptions on the data and failing out incorrectly.
When an error occurs either a valid error payload is returned or something else is returned.
Do not be opinionated about errors payloads here because the fetchUser should not care at this time.

Preserve the original user role and if it is not defined then set the role to ROLE_ANONYMOUS.
Only when the payload has a valid Credentials structure should the session role be deleted and the credentials object be modified.

Having a token when the payload is invalid or the role is anonymous  results in an error message getting stuck.
Delete the token in the case that the role is anonymous.
After all, the anonymous user should not have a token.

Continue to perform the completion steps needed to resolve the fetchUser request regardless of the payload structure.