Security fixes for: minimatch and qs.
Fix child dependencies that are coming in with insecure versions.
This is based on the `npm list --depth=100` and considers what is reported by dependabot for Sage.

There is a concern with `minimatch`.
One of the dependencies is trying to pull in a `5.1.0` version (or higher).

see: https://github.com/TAMULib/SAGE/security/dependabot/37
see: https://github.com/TAMULib/SAGE/security/dependabot/34