Issue 221: Return to Safety, results in going back to 403.
This is happening because of a double redirect.
When going to the `/management` page, a browser is redirected to `/management/internal-metadata`.
Then the last route ends up referring to the `/management` page, which is the route path of `/management/:tab`.
The `:tab` is not a valid part of the URL because the `$scope.lastRoute` is expected to be a path rather than a route.
Rather than rename `lastRoute` to `lastPath`:
1. Add a comment describing this behavior.
2. Wrap the assignment in a function that changes the path if any route parts are found.