Weaver Repo Lazy is too Eager.
The weaver repo is always loading data, even if already loaded, when lazy is set to TRUE.

Lazy is being set to TRUE for performance improvements and for granting control when something is loaded (if at all).
But when lazy is set to TRUE, it is always loading the data rather than using the existing cached data.
This has a potentially huge performance hit for large data sets on pages depending on the circumstances.