Provide an alternative to getAll() that returns a promise and still respects the lazy load setting.
Use `findAll()` for this purpose.
This better allows for the use-case where the lazy-load is loaded right before it is needed.
The current popularized design is to load everything on page view, even when lazy loaded.

The promise callback being returned is necessary so that the loaded values can be processed after successful load.